# api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: "0.0.1"
  title: BugTracktor
  description: A flexible bug tracker.
  contact:
    name: Alex Isaienko
    email: thesonerik@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: bug-tracktor.herokuapp.com
basePath: /v1
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json


## Security
securityDefinitions:
  api_key:
    type: apiKey
    name: api_key
    in: header
security:
  - api_key: []

## Cross-request parameters
parameters:
  projectIdParam:
    in: path
    name: projectId
    description: ID of the project to operate with.
    required: true
    type: integer
    format: int64
...
  issueIndexParam:
    in: path
    name: issueIndex
    description: Index of the issue to operate with.
    required: true
    type: integer
    format: int64
  fromDateQueryParam:
    in: query
    name: fromDate
    description: Minimum date of the returned items.
    required: false
    type: string
    format: date
  toDateQueryParam:
    in: query
    name: toDate
    description: Maximum date of the returned items.
    required: false
    type: string
    format: date
  limitQueryParam:
    in: query
    name: limit
    description: Max size of the returned array.
    required: false
    type: integer
    format: int32
...
paths:
...
  /projects/{projectId}/issues:
    x-swagger-router-controller: issues
    parameters:
      - $ref: "#/parameters/projectIdParam"
    get:
      operationId: listIssues
      description: |
        Gets a list of issues for a given project.
      parameters:
        - $ref: "#/parameters/limitQueryParam"
        - $ref: "#/parameters/fromDateQueryParam"
        - $ref: "#/parameters/toDateQueryParam"
      responses:
        200:
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/Issue"
    post:
      operationId: createIssue
      description: |
        Creates a new issue.
      parameters:
        - in: body
          name: issue
          description: The issue to create.
          required: true
          schema:
            $ref: '#/definitions/Issue'
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/Issue"
  /projects/{projectId}/issues/{issueIndex}:
    x-swagger-router-controller: issues
    parameters:
      - $ref: "#/parameters/projectIdParam"
      - $ref: "#/parameters/issueIndexParam"
    get:
      operationId: getIssue
      description: |
        Gets an issue with a given identifier.
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/Issue"
    patch:
      operationId: updateIssue
      description: |
        Updates an issue with a new data (should produce a new IssueChange).
      parameters:
        - in: body
          name: issue
          description: The updated issue.
          required: true
          schema:
            $ref: "#/definitions/Issue"
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/Issue"
  /projects/{projectId}/issues/{issueIndex}/changes:
    x-swagger-router-controller: issues
    parameters:
      - $ref: "#/parameters/projectIdParam"
      - $ref: "#/parameters/issueIndexParam"
      - $ref: "#/parameters/limitQueryParam"
      - $ref: "#/parameters/fromDateQueryParam"
      - $ref: "#/parameters/toDateQueryParam"
    get:
      operationId: getIssueChanges
      description: |
        Gets changes of the given issue.
      responses:
        200:
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/IssueChange"
...
definitions:
  Issue:
    type: object
    properties:
      id:
        type: integer
        description: Unique issue identifier.
      issue_index:
        type: integer
        description: Index of the issue inside the project.
      is_opened:
        type: boolean
        description: Whether the issue is opened for discussion.
      is_active:
        type: boolean
        description: Whether the issue should be visible for project members.
      author:
        $ref: "#/definitions/User"
      assignees:
        type: array
        items:
          $ref: "#/definitions/User"
      attachments:
        type: array
        items:
          $ref: "#/definitions/IssueAttachment"
      project:
        $ref: "#/definitions/Project"
      type:
        $ref: "#/definitions/IssueType"
      history:
        type: array
        items:
          $ref: "#/definitions/IssueChange"
      short_description:
        type: string
        description: Short description of the issue.
      full_description:
        type: string
        description: Full description of the issue.
      creation_date:
        type: string
        format: dateTime
        description: Date when issue was created.
  Project:
    type: object
    properties:
      id:
        type: integer
        description: Unique project identifier.
      name:
        type: string
        description: Project name.
      short_description:
        type: string
        description: Short description of the project.
      full_description:
        type: string
        description: Full description of the project.
      creator:
        $ref: "#/definitions/User"
      members:
        type: array
        items:
          $ref: "#/definitions/ProjectMember"
      issues:
        type: array
        items:
          $ref: "#/definitions/Issue"
  User:
    type: object
    properties:
      id:
        type: integer
        description: Unique user identifier.
      email:
        type: string
        format: email
        description: User's email.
      password:
        type: string
        description: User's password.
      nickname:
        type: string
        format: nickname
        description: User's nickname.
      real_name:
        type: string
        description: User's real name.
      avatar_url:
        type: string
        description: User's avatar URL.
  IssueType:
    type: object
    properties:
      id:
        type: integer
        description: Unique issue type identifier.
      name:
        type: string
        description: Issue type name.
      description:
        type: string
        description: Issue type description.
  IssueChange:
    type: object
    properties:
      issue_id:
        type: integer
      author_id:
        type: integer
      date:
        type: string
        format: dateTime
      change:
        type: string
  Role:
    type: object
    properties:
      id:
        type: integer
        description: Unique role identifier.
      name:
        type: string
        description: Role name.
      description:
        type: string
        description: Role description.
      permissions:
        type: array
        items:
          $ref: "#/definitions/Permission"
  Permission:
    type: object
    properties:
      name:
        type: string
        description: Unique permission name.
      description:
        type: string
        description: Permission description.
  IssueAttachment:
    type: object
    properties:
      issue_id:
        type: integer
        description: Unique issue id this attachment belongs to.
      url:
        type: string
        description: Attachment url.
  ProjectMember:
    type: object
    properties:
      user:
        $ref: "#/definitions/User"
        description: User entity that is attached to the project as it's member.
      project:
        $ref: "#/definitions/Project"
      roles:
        type: array
        items:
          $ref: '#/definitions/Role'
      join_date:
        type: string
        format: dateTime
      exit_date:
        type: string
        format: dateTime
...